Get mpd duration as ExoPlayer event listener
Rather than take a reference to the player and query that every time we
need the duration, this setup means the sample processor will be
updated by the player every time there’s a new manifest, and can grab
the value then.

I prefer this setup because we don’t have to co-ordinate the creation
of the player and the sample processor as much, and we won’t have the
sample processor receiving the reference to the player after it’s
constructed – that was bugging me because for a short space of time the
sample processor doesn’t have all the info it needs.

I’ve also included a constant in the sample processor to mark when data
is unavailable, in case the sample processor doesn’t receive the data
before it’s requested.