Write switchable sampler
This sampler can be switched between a couple of different modes
(including time- and size-based estimation), and should be able to
replace both the TimeBasedSampler and the SizeBasedSampler.

At the moment the class doesn’t have a mode for chunk-based sampling –
I’ll need to have a look at that.