Make logger get buffer value directly from player
Previously, the logger was being given its buffer estimate by the chunk
source, but I was unhappy with that because that didn’t feel like it
should be a responsibility of the chunk source.

I’ve changed it so the logger gets the buffer level by querying the
player directly, using the getBufferedPosition() and
getCurrentPosition() methods.

However, these values are different. I think the buffer value now is
the value *after* the chunk has been downloaded, whereas before it was
the value before the chunk has been downloaded. We’ll need to revert
this if that matters.