Make sure actual rate is in kbps, and round rates to integer value
For the logging, we don’t need the fraction and we want all rates to be
in kb/s.

Even though actual rate was labelled as kbps, it was actually in bits
per second, so I’ve fixed that, and I’m now rounding the actual rate
and delivery rate values to the closest integer.