Revert "Make logger get buffer value directly from player"
This reverts commit 73f66f80b9d24a9aacfe96831ed4e0f56c66b0d0.

For logging the buffer level, we need to be logging the value that was
correct when the segment had finished downloading.

Getting the value directly from the player didn’t accomplish this, as
the buffer level was being calculated when the log was being created,
which was some time later.