Pass previous chunk with buffer estimate to prevent duplicates
As the chunk source method getNextChunk() is currently being called
twice for the first 9 chunks (still not sure why), the buffer level is
being updated multiple times.

We only want the buffer level to be updated once for each chunk, and
I’m going with the first time. By passing the chunk with the buffer
update, we can compare it to the last received chunk and only accept
the updated buffer level if the two chunks are different.