Synchronise buffer update with onLoadCompleted()
Previously, the buffer updates and the onLoadCompleted() callback were
not co-ordinated – it was assumed that the buffer level would be
updated before onLoadCompleted().

However, this is not generally correct – when running, I found that
onLoadCompleted() fired before the buffer level update for the first
few downloaded chunks, and after that they inverted and the buffer
level was updated first.

To fix this, I’ve implemented a basic latch, where the log entry isn’t
created until there’s a new buffer level value and also new chunk data.

The buffer levels being written to the log now make sense.