Log chunk arrival times relative to the manifest request time
Before now, chunk arrival times were given as the value of
SystemClock.elapsedRealtime().

However, it’s better for those times to have a constant reference point
across runs, so that they can be easily compared.

As a reference point, the beginning of the manifest request makes
sense, as the resulting time shows any processing delays between
beginning that request and requesting the first chunk.