Wait for stall information before logging
Before now, the stall-ending callback (onPlayerStateChanged()) was
being called too late, coming after the buffer update and the chunk
data. This meant that the stalls were recorded one chunk too late.

Now, when the buffer and chunk info come in, we’re checking whether a
stall occurred this time (and the stall info hasn’t been delivered
yet). If so, we wait for the stall information to come in before
creating the log.

This should also work if the stall information comes in before either
of the others.