Display helper info onscreen
We want to display information like the current buffer level onscreen,
to make it easier to tell what’s going on. Quentin seems to have had
this working with the getInfosVideo() method.

The method was largely commented out, and wasn’t being called. Based on
the variable names, I think this was because it wasn’t working
correctly.

I have uncommented it and deleted a lot of the information, as it’s
either not useful or I don’t know what it is. I’ve also added a bunch
of null checks to prevent null pointer exceptions.