Improve (and rename) getInfosVideo()
Rather than build the strings by concatenation, I’ve started using a
StringBuilder with a Formatter, as then I don’t have to create extra
string variables.

I’ve also added an extra null check as the player might be null if this
function is called when the video is being exited.

To make the function clearer, I’ve renamed it and added javadoc.