Make time-based sampler using Handler
Before now, the time-based sampler checked the time every time
onBytesTransferred() was called, and delivered a sample if the time
threshold had been exceeded.

The problem with this is that if the network rate becomes low enough
that onBytesTransferred() isn’t being called often enough, the samples
will no longer be made frequently enough. This means the sampling can’t
be used to detect a sudden drop in network rate while downloading a
particular chunk, for example.

As an alternative, I have used a Handler to schedule sampling to happen
every x ms once a transfer has started. There will also be a sample
made when the transfer finished.

I have chosen Handler over Timer and TimerTask, as it’s apparently more
reliable.