Update debug view text rather than appending to it
As the StringBuilder is an instance variable, adding new text to it
every time we needed to update the debug view was causing the view to
fill up with text that flowed off the screen.

To prevent this, I’ve added a line to clear the text in the
StringBuilder each time the update function is called.