Use separate clocks for time samples and chunk samples
Before now, when a transfer finished a sample was delivered covering
the time since the last delivered sample. So if there had been
time-based samples in the middle of downloading a particular chunk, the
final sample would be only for the time between the last time-based
sample and the end of the transfer.

Rather than doing this, I’ve added a separate clock variable for the
transfer-end samples, so that they will cover the entire transfer.

In this way, when a chunk download is taking a long time relative to
the one before it, we’ll get additional time-based samples in the
meantime that will give us an indication of the network rate, and allow
us to adapt before the chunk has finished downloading. Then, when the
chunk does finally download, we’ll have an average throughput for the
whole chunk as well.