Compose chunk sampler inside time sampler
Before now, the time sampler was creating a sample at the end of every
transfer. Some of those transfers weren’t media chunks though –
initialisation transfers and quality changes.

To make sure that there’s only either time-based samples or chunk
samples, I’ve composed an instance of ChunkBasedSampler inside the
time-based sampler. When a chunk sample is complete, the chunk-based
sampler sends it to the time-based sampler, which can then send it on
to the sample store.

The time-based sampler needs to be notified of the chunk samples so
that it can stop the handler that’s performing the time samples.

I suspect composition like this will be useful in creating different
samplers, so I think I should rename this time-based sampler to
indicate that it’s also performing chunk sampling, and create a
separate purely time-based sampler.