Write sample logs to file
Before now, we’ve been writing the chunk log to file after each run,
but keeping the samples internally.

For time-based sampling where we have both time samples and chunk
samples, it’s useful to be able to compare logs for the samples with
the log for the chunks, in order to evaluate how the sampling is
working.