Use dedicated manifest listener to get the request time
Before now, the chunk logger was registered as a listener on the
manifest data source so it could get the manifest request time. This
time is used as the reference point for the chunk arrival times in the
log.

Since we want to create a log of the samples as well to compare, the
arrival times of samples will need to use the same reference point (the
manifest request time).

So I’ve created a dedicated manifest listener, on which both the chunk
logger and the sample processor can be registered as listeners. At the
moment, the sample processor doesn’t do anything with the manifest
request time, as it’s not logging the arrival times of the samples.

In the ManifestListener, I’m not sure if the notifying of listeners
should occur on a background thread or not. For the moment I’ve left it
on the same thread, but I’m slightly worried this could eventually
become too much processing.