Don't extend listener interfaces with ChunkLogger
Before now, the ChunkLogger interface was extending the listener
interfaces (e.g. TransferListener) that the DefaultChunkLogger was
using to get its information.

I think that these interfaces represent the implementation of the chunk
logger by the DefaultChunkLogger class and so should be implemented by
that class rather than extended by the ChunkLogger interface.