Include sample arrival time in sample log
To allow comparison with the chunk log, I’m including the arrival time
of the samples when storing them.

At the moment the arrival time is calculated from the clock when the
samples are received by the processor – this value will be slightly
larger than it should be. During a test run, I found it to be out by ~1
ms.