Calculate sample arrival time as early as possible
Rather than being calculated when the sample processor receives the
sample, the sample arrival time is now calculated when the sample is
made. This ensures that the sample arrival times are accurate.

I’ve kept the manifest request time offset with the sample processor,
as I don’t think the individual samplers should have to deal with that.