Remove divisor of 1000 from static parameter calculation
I’m just doing this on a hunch rather than anything else, but I think
that factor of 1000 is a conversion factor for bps/kbps.

Once removed, the algorithm updates quite quickly to the maximum
quality, though not as quickly as before changing any of the units. I
also tested how the algorithm responds to lower network rates, and it
responds well, decreasing effectively. I’m not sure if it was
responding correctly before.