Use clearer variable names in OscarHTrackSelection
Some of the variable names were very unclear, so to make the code
easier to understand, I’m renaming them.

The variables m_staticAlgPar and m_staticAlgPar2 had the names they had
because they needed to be shared across algorithms in a previous
implementation, and so the names needed to be very general.

As this is no longer the case, I can now rename them. Ahmed suggested
the names “accumulatedSwitchingCredit” and “switchingCreditUnit” – the
variables are used to determine whether switching occurs and how much
it occurs by.