Make chunk sampler send samples to time sampler
The time-based sampler was supposed to be notified when chunk samples
were finished, so that it could stop making time-based samples.

However, I had mistakenly passed the sample store to the chunk sampler
instead of passing the time sampler to it. This meant that the chunk
sampler was notifying the sample store directly, and the time sampler’s
addSample() method was not being called.

I have fixed this now by passing the time sampler as the chunk
sampler’s sample store.

Note that this commit hasn’t been tested, as I found the bug after
making other changes, which I’m choosing to commit out of order.