Don't deliver a chunk sample if time sampling has started
Time sampling starts when a chunk takes too long to download. When the
slow chunk finally arrives, the chunk sample for that chunk will cover
the same period of time as the time samples, which is undesirable.

Rather than do this, when a chunk finishes and time samples have been
delivered, the time based sampler now delivers a time sample covering
the period of time between the last full time sample and the delivery
of the chunk, and doesn’t deliver the chunk sample.