Change max level switch back to 2
This constant determines how many representation levels the algorithm
is allowed to increase by. It was at 10, but the original algorithm
code has the value 2.

As the algorithm is meant to only incrementally increase, I’ve changed
it back to 2.

However, the quality level still increases very slowly, so this is not
what was causing the problem.