Divide switching credit unit by (number of tracks - 1)
In the original version of the algorithm, this division is by (number
of tracks - 1), whereas at some point it has become dividing by the
number of tracks.

I reckon this is a change I introduced when replacing how some things
were calculated, but I doubt it’s in one commit on its own, so I’m not
reverting.