Simplify buffer threshold comparisons
The algorithm makes some comparisons between the current buffer
occupancy and the total buffer occupancy. In the original C++ code, the
maximum time length of the buffer is found by multiplying the number of
chunks by the chunk duration. However, since we already have that value
in time units, there’s no need to convert it to number of chunks and
then convert it back.

There was also a factor of 1000 that shouldn’t have been there – I
suspect I incorporated this when changing how the calculations were
done.