Provide last chunk bitrate in bps for compatibility
In ExoPlayer, the track formats have their bitrate in bps. As we were
previously handling all rates internally in kbps, getRepIndex() used a
kbps value to find the index, which was incompatible with the track
formats as they are available to the track selections.

For consistency, I’ve changed the sample processor to report the last
representation level in bits per second instead of kilobits per second,
which means the new getRepIndex() now functions correctly.