Remove redundant class SwitchableSampler
This class was created to offer switching between size-, time-, and
chunk-based sampling, as they seemed very similar.

However, the time-based sampling method that this sampler uses is not
useful, and we are no longer using it. The newer method is not as
compatible with size-based sampling, and switching will be more
complicated.

Also, the TimeBasedSampler uses chunk based sampling unless a chunk
takes too long, and so covers one application of switchable sampling.