Use timer to delay requests
Before now, the MISLLoadControl didn’t have any custom logic for
determining when to continue loading, and just followed the default
ExoPlayer behaviour.

I’ve added an extra delay, so that requests won’t happen more
frequently than that delay.

One possible issue with this implementation is that it assumes the
player will keep calling shouldContinueLoading() relatively regularly –
I haven’t looked at the details of that yet, but so far it seems to
work.