Use Handler for updating debug view
Before now, the code created a background thread, which would call back
into the UI thread at a fixed interval to update the debug view.

I’ve replaced this code with posting a runnable to the handler on the
UI thread. The resulting code is shorter and less complicated, and
achieves the same thing without the background thread. It may mean that
the timing of the debug view updating is slightly less guaranteed, but
I don’t think that matters too much.

I’ve also changed the debug view to update every 1000 ms instead of
1500 ms, as I found the interval slightly too slow.