Ignore repeated requests for the same chunk
When the algorithm was called more than once for the same chunk, it
would sometimes initially suggest switching up in quality, and then
suggest remaining at the same quality next time. I think this is
because the algorithm assumes that the switch up happened, and updates
the switching credit accordingly (and wrongly).

I’ve changed the track selection to only run the algorithm if a new
chunk has been downloaded since it was last called.