Simplify arbiter calculations
The buffer occupancy was being calculated by converting the times into
numbers of chunks and then getting their ratio – I’ve removed the
number of chunks conversion, as the ratio should be the same.

Also, the index for the last rate was being found using the
representation rate for the last chunk, which will make the algorithm
harder to use with other forms of sampling.

As an alternative, I’m using the last value of the field selectedIndex.
This will work as long as the algorithm isn’t called multiple times per
chunk, changing selectedIndex each time.

Guaranteeing that it’s not called more than once per *sample* (more
generally than chunk) should be pretty straightforward, which would
more easily allow the algorithm to work with other sampling methods.