Add method for checking for decreasing throughput
Arbiter+ needs to check for decreasing throughput as part of its sudden
drop detection.

In the C++ this was implemented as comparing the two most recent
samples directly in the algorithm, but I’ve moved it to the sample
processor – I think it’s beneficial to maintain as much separation
between the algorithms and the samples as possible, to allow
flexibility in the sample storage.