Create Arbiter+ track selection
This track selection implements the Arbiter+ algorithm.

At the moment, this code doesn’t work, as there are a few things that
need to be fixed/figured out.

The original code expects the tracks to be ordered the other way around
(lowest quality at index 0), so the track-related logic needs to be
tweaked. I have already made some changes as I transcribed, which I
probably shouldn’t have done.

Also, quality scores for future chunks aren’t available yet, so that
part of the algorithm is unimplemented.