Use the last value of selectedIndex in Arbiter+
Before now, to get the last selected index, it was found from
sampleProcessor.lastChunkRepLevel().

However, using the last value of selectedIndex should work, and won’t
require a call to the sample processor, which means less is required
from the sample processor.

One issue is if the algorithm is called multiple times per chunk and
changes this value. However, it should be possible to prevent this.