Use builder pattern in Arbiter+ creation
Before now, the approach with track selections was that you could
either use all the default values, or specify every parameter.

To allow just some values to be changed from their defaults, I’ve used
a basic version of the builder pattern when creating the
ArbiterPlusTrackSelection factory. Now any of the parameters as desired
can be left as default values or changed.

However, I’m not sure how useful this will be in practice. I suspect
when configuring a track selection we will always be specifying all of
the parameters in a configuration file, in which case using the builder
pattern as I have just adds unnecessary code.