Only calculate chunkSizeRatio if detectSuddenDrop is set
As the chunk size ratio is only used by the sudden drop detection code,
it doesn’t need to be calculated if sudden drop detection is not being
used.

Also, this clears up some uncertainty about using Arbiter+ with
time-based sampling – as long as we keep sudden drop detection off (and
it isn’t needed when we have time-based sampling anyway), the algorithm
works fine with time-based sampling.