Get buffer level from player instead of chunk source
Providing the chunk logger with buffer estimates by using the
replacement for the DefaultDashChunkSource means that the logging won’t
work for HLS and SmoothStreaming.

We can make it work for SmoothStreaming by replacing
DefaultSsChunkSource as with DefaultDashChunkSource, but this approach
won’t work for HLS, as HlsChunkSource is a package-private class that
we can’t replace or extend.

What I’ve decided to go with is calculating the current buffer level
when onLoadCompleted() is called by requesting info from the player.
These values are slightly lower (40 ms – 120 ms) than the estimates
provided to the track selections, which suggests that the values are
slightly inaccurate, though they are consistent with the other values
the logger receives.

One problem that will have to be fixed is that the buffer value
recorded for the first chunk is incorrect – the player doesn’t update
its buffered position value completely before onLoadCompleted() is
called, so the calculated buffer value can range from 0 ms to 4000 ms.