Check for media data rather than video track in logger
Before now, we were checking that the load which had been completed was
from a video track.

With HLS, this doesn’t work, as the video and audio are demuxed, and so
the track type value is always a default track value.

Checking for the data type to be media seems to work instead for HLS,
though it hasn’t been tested with DASH yet.