Add type check before casting manifest
The manifest duration is obtained by casting the manifest (which is
received as an Object) to DashManifest, which causes a runtime error
when we’re using HLS.

I’ve added a type check against DashManifest and HlsManifest, and set
the mpd duration accordingly for each. I’ve also added a check and cast
for SsManifest, as I might as well do it now (though it is untested).