Use TrackSelection indices not TrackGroup indices
I had thought that the track selections should be setting selectedIndex
according to the indices specified in the tracks array. However, it
appears that we should be setting it according to the sorted order the
track selections use.

For our video, the tracks are already sorted in the correct order, so I
didn’t notice this mistake before – it came up with the HLS video,
which isn’t in sorted order.