Tidy up DefaultSampleProcessor
I’ve removed all javadoc that was copied from implemented interfaces. I
reckon this duplication is going to lead to conflicting documentation,
as it’s easy to forget to update the documentation in all implementing
classes.

I’ve also reworded a couple of javadoc comments to match the style in
SampleProcessor, and removed redundant words in method names as I did
there.

I’ve also changed the average and variance static methods to have a
public visibility. I think those methods may be useful to other classes
implementing SampleProcessor, and as they don’t depend on any of the
class’ implementation, I don’t think it’ll be a problem for other
classes to depend on them.

Ideally I would like to put those methods in the SampleProcessor
interface, but that’s not possible below android nougat.