Remove getNearestBitrateIndex()
This method is extremely similar to findBestRateIndex(), but selects
the lowest bitrate above a target instead of the highest below a
target. I’d like to get rid of it and depend only on
findBestRateIndex().

The only algorithm still using this method is BBA2, so I’m changing it
to use the other here.