Don't provide option to set net scaling and buffer factors
I had provided an option to set initial values for these fields, but I
don’t think that should be possible – the initial value for the
buffering factor is never used, and the net scaling code seems to
expect the net scaling factor to start at 1.

I have now removed this option, and have just given initial values to
the two parameters.