Remove time-based estimation window parameter
The time-based estimation window parameter wasn’t being used for
anything. In the original C++ it was used if the estimation mode was
currently set to time-based estimation, but there is no estimation mode
here.

If we want to set it up to consider a different number of samples when
using time-based estimation, we will need to either modify the sample
processor to track whether samples are time-based or not, or track the
sampling mode within the algorithm.

For the moment, I’ve removed the time-based estimation window parameter.