Move sample processor to track selection superclass
As the sample processor is common to most of the algorithm track
selections, I’ve moved it to the superclass and made it a protected
member.

Note that this is a bit weird for DASHTrackSelection, which doesn’t use
a sample processor at all – at the moment I’m passing null to the super
constructor, which is a little strange.

This change should make it easier to move common functionality to the
superclass as well, as the sample processor will now be available there.