Move OscarKumarParEstimation back to OscarHTrackSelection
This method is an estimation calculation that is very particular to
OscarHTrackSelection, and it’s not particularly reusable, so I don’t
think it’s worth having it in the SampleProcessor interface or
implementation.

It might be useful to break the method up into separate bits, and move
any of those that are reusable to the sample processor, but that would
be a pretty big job, and would require a clearer understanding of the
calculations involved than I have.

One consequence of moving this method back to OscarHTrackSelection is
that I had to make throughputSamples() and windowSize() public again. I
think this is ok – the algorithms are still independent of how the
throughput samples are stored (e.g. of the array indices).