Rework FutureSegmentInfos to store information directly
Before now, FutureSegmentInfos was a class for storing the information
on one specific video segment, and the overall information was stored
in an ArrayList elsewhere in the code.

One significant problem with this approach was that to find the size of
a chunk, a loop was required through the list storing the information.

I’ve reworked the class (and renamed it, to be consistent with the use
of “chunk” over “segment”) to instead store all of the information in
one instance, and provide access to it through an instance method.

The class uses an arrangement internally that gives O(1) amortised
lookup cost for the byte size. For ease of coding, I’ve used HashMaps,
even though they’re filled completely in an array-like fashion (they
aren’t necessarily filled in order, and working out the pre-sizing of
the arrays is awkward). It should be easy to convert to lists in future.