Incorporate FutureChunkInfo into code
Following from the last commit, I’ve replaced use of the list of
FutureSegmentInfos instances with an instance of FutureChunkInfo that
stores all the information.

Queries should be faster, and the method calls are simpler (instead of
calling a static method and passing it the relevant array, I can call
the method directly on a static reference to the class instance).