Remove redundant chunk duration calculations
In a couple of places in BBA2, the reservoir (which is in ms) is
compared to the maximum buffer length in ms.

However, this maximum buffer length is found by dividing the maximum
buffer length in ms by the chunk duration in ms to give the maximum
buffer length in number of chunks, and then multiplying that value by
the chunk duration to convert it back to ms.

As far as I can tell, the conversion is completely redundant and
unnecessary, and I think it came about as a result of switching from
GPAC to ExoPlayer – in GPAC I think the buffer size is given as a
number of chunks.