Fix window calculation for checking future chunk sizes
I think this calculation had an error in it – now that I’ve changed it,
it’s setting the window to twice the maximum number of chunks in the
buffer, or however many chunks are left in the video if there aren’t
that many.

I can’t make sense of what the calculation might have meant the way it
was. I’m pretty sure this was an error.