Create LogBuilder class
This class will provide the interface for building different logs.

I’ve used an abstract class rather than an interface in order to
provide default empty definitions for the methods. This way, subclasses
can override only the methods they need, and more methods can be added
to the LogBuilder later.