Use LogBuilder in DefaultChunkLogger
Rather than code the log-writing directly into the DefaultChunkLogger
class, delegating to another object will allow the log to be changed by
providing a different object, which should make it easier to change the
format of the logging.

At the moment, the log builder is instantiated directly, but the next
step is to allow it to be passed in at creation.

This works, and produces a log as expected.

One possible issue is that it is a bit verbose specifying each log
value separately, but I think that’s not so bad.

I’ve removed the toString() method on the log entries, that was being
used to create the line in the log file – it’s not needed anymore.

I’ve also removed the getter/setter methods. I don’t think it’s too big
a deal to depend on the fields directly within this class – the private
visibility limits how wide that dependency is, and I don’t see us
wanting to change how this access works. I’ve also changed the fields
to be final, as I don’t think they should need to change.