Let DefaultChunkLogger's log builder be changed
By letting the log builder be passed in, we should get a lot of
flexibility for changing the log format in the future.

I’ve also included a constructor that just takes a file and uses the
default log builder with that file, and moved the creation of the file
name (using the current time) to the PlayerActivity – I think these are
hardcoded configuration details that should be obvious at a high level,
so that they can be changed/be configurable.