Let DefaultSampleProcessor's log builder be changed
Allowing the builder to be passed in as a parameter should make the log
structure and format more flexible.

This required creating the File object for the sample log before
configureRun() is called in the PlayerActivity, so I moved that method
call from onCreate() to initializePlayer(). I think this actually makes
more sense as a place for this method anyway – if we’re doing multiple
runs without reloading the app, this is when the configuration should
be done.

I also implemented the throughput() method in DefaultLogBuilder,
because I had forgotten to do that. I think possibly it would be better
to use an interface for LogBuilder instead of an abstract class, and
just override irrelevant methods with empty methods in implementing
classes.