Move request scheduling logic to a new class
I’m going to be merging this into the master branch, as I’ll only be
able to merge a single branch into the MISL repository and probably the
AT&T repository as well.

I want this to stay as an example of custom request scheduling, so I’m
putting it in a separate class so it won’t be overridden by later
updates to MISLLoadControl.