Remove Unpublished Algorithms
I would like to be able to display this project as an example of the
work I’ve done over the summer, but these three algorithms are
unpublished, so I need to remove them before making this project public.

After this commit I’m going to remove the files from the project
history using the bfg tool.